+++ src/calendarjs.js
@@ -61,6 +61,7 @@
  * @property    {Object}    onEventsExported                            Specifies an event that will be triggered when the "Export Events" button is pressed.
  * @property    {Object}    onSetDate                                   Specifies an event that will be triggered when the date on the main display is set externally.
  * @property    {Object}    onEventsSet                                 Specifies an event that will be triggered when events are set and the originals are cleared (passes the events to the function).
+ * @property    {Object}    onGroupsCleared                             Specifies an event that will be triggered when the event groups are cleared.
  */
 
 
@@ -5061,6 +5062,48 @@ function calendarJs( id, options, startDateTime ) {
     function toStorageDate( date ) {
         return date.getFullYear() + "-" + date.getMonth() + "-" + date.getDate();
     }
+
+
+    /*
+     * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+     * Get/Set Additional Data (public)
+     * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+     */
+
+    /**
+     * getAllGroups().
+     * 
+     * Returns an array of group names being used.
+     * 
+     * @returns     {Object[]}                                              An array of the group names.
+     */
+    this.getAllGroups = function() {
+        return getGroups();
+    };
+
+    /**
+     * clearAllGroups().
+     * 
+     * Clears all the event groups.
+     * 
+     * @fires       onGroupsCleared
+     * 
+     * @param       {boolean}   updateEvents                                States of the calendar display should be updated (defaults to true).
+     */
+    this.clearAllGroups = function( updateEvents ) {
+        updateEvents = !isDefined( updateEvents ) ? true : updateEvents;
+
+        getAllEventsFunc( function( event ) {
+            event.group = null;
+        } );
+
+        triggerOptionsEvent( "onGroupsCleared" );
+
+        if ( updateEvents ) {
+            buildDayEvents();
+            refreshOpenedViews();
+        }
+    };
 
 
     /*